split up ad prefs.js
