+++ static/src/javascripts/projects/common/modules/identity/ad-prefs/BoopableBox.js
@@ -1,0 +1,37 @@
+import React, { Component } from 'react';
+
+class BoopableBox extends Component<
+    BoopableBoxProps,
+    {
+        booping: boolean,
+    }
+    > {
+
+    constructor(props:BoopableBoxProps) {
+        super(props)
+        this.state = {
+            booping: false
+        };
+    }
+
+    boop() {
+        this.setState({ booping: true });
+        setTimeout(() => {
+            this.setState({ booping: false });
+        }, 500);
+    }
+
+    render() {
+        return (
+            <div
+                className="identity-ad-prefs-manager__boop"
+                style={{
+                    display: this.state.booping ? 'block' : 'none',
+                }}>
+                {this.props.children}
+            </div>
+        );
+    }
+}
+
+export {BoopableBox}
